portfolio/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
alembic
pydantic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
redis
faiss-cpu
langchain
langchain-community
python-dotenv
httpx
google-generativeai
sentence-transformers
supabase
